fixed transaction to show 2 decimal places

+++ models/transaction.rb
@@ -7,7 +7,7 @@ class Transaction
 
   def initialize( transaction )
     @id = transaction["id"].to_i if transaction["id"]
-    @amount = transaction["amount"].to_i
+    @amount = transaction["amount"].to_f
     @tag_id = transaction["tag_id"].to_i
     @merchant_id = transaction["merchant_id"].to_i
   end
